+++ doctor-plugin/src/main/java/com/osacky/doctor/DoctorExtension.kt
@@ -23,6 +23,10 @@ open class DoctorExtension(objects: ObjectFactory) {
      */
     val GCWarningThreshold = objects.property<Float>().convention(0.10f)
     /**
+     * The level at which to fail when a build spends more than this percent garbage collecting.
+     */
+    val GCFailThreshold = objects.property<Float>().convention(0.9f)
+    /**
      * Print a warning to the console if we spend more than this amount of time with Dagger annotation processors.
      */
     val daggerThreshold = objects.property<Int>().convention(5000)